Move the internal package into context
