+++ opentelemetry/src/main/scala/zio/telemetry/opentelemetry/tracing/propagation/TraceContextPropagator.scala
@@ -2,7 +2,7 @@ package zio.telemetry.opentelemetry.tracing.propagation
 
 import io.opentelemetry.api.trace.propagation.W3CTraceContextPropagator
 import io.opentelemetry.context.propagation.TextMapPropagator
-import zio.telemetry.opentelemetry.internal.Propagator
+import zio.telemetry.opentelemetry.context.internal.Propagator
 
 trait TraceContextPropagator extends Propagator
